+++ R/coor.test.R
@@ -21,8 +21,16 @@
 #' @param rm.solo Logical. Controls if signals that are not intercalated at the start or end of the 
 #' sequence are removed (if \code{TRUE}). For instances the sequence of signals A-A-A-B-A-B-A-B-B-B (in which A and B represent different individuals, as in the 'indiv' column) would be subset to 
 #' A-B-A-B-A-B. Default is  \code{FALSE}.
-#' @return A data frame with the observed number of overlaps (obs.overlaps), mean number of overlaps expected by chance,
-#' and p value.  
+#' @return A data frame with the following columns:
+#' #' \itemize{
+#'    \item \code{sing.event}: singing event ID
+#'    \item \code{obs.overlaps}: observed number of overlaps
+#'    \item \code{mean.random.ovlps}: mean number of overlaps expected by chance
+#'    \item \code{p.value}: p value 
+#'    \item \code{coor.score}: coordination score (**sensu** Araya-Salas et al. 2017), 
+#'    calculated as `(obs.overlaps - mean.random.ovlps) / mean.random.ovlps`. 
+#'    Positive values indicate a tendency to overlap while negative values indicate a tendency to alternate.
+#'    }
 #' @export
 #' @name coor.test
 #' @details This function calculates the probability of finding and equal or lower number 
@@ -32,7 +40,15 @@
 #' expected values. The p-values are calculated as the proportion of random expected values that were lower (or higher) 
 #' than the observed value. The function runs one test for each singing event in the input data frame. The function 
 #' is equivalent to the "KeepGaps" methods described in Masco et al. 2015.
-#' @references Masco, C., Allesina, S., Mennill, D. J., and Pruett-Jones, S. (2015). The Song Overlap Null model Generator (SONG): a new tool for distinguishing between random and non-random song overlap. Bioacoustics. 1-12. 
+#' @references 
+#' {
+#' Araya-Salas M., Wojczulanis-Jakubas K., Phillips E.M., Mennill D.J., Wright T.F.\
+#'  (2017) To overlap or not to overlap: context-dependent coordinated singing in 
+#'  lekking long-billed hermits. Anim Behav.
+#' Masco, C., Allesina, S., Mennill, D. J., and Pruett-Jones, S. (2015). The Song 
+#' Overlap Null model Generator (SONG): a new tool for distinguishing between random
+#' and non-random song overlap. Bioacoustics.
+#' } 
 #' @examples{
 #' #load  simulated singing data (see data documentation)
 #' data(sim.coor.sing)
@@ -44,7 +60,7 @@
 #' coor.test(sim.coor.sing, iterations = 100, less.than.chance = FALSE)
 #' }
 #' @author Marcelo Araya-Salas (\email{araya-salas@@cornell.edu})
-#last modification on jul-5-2016 (MAS)
+#last modification on apr-11-2018 (MAS)
 
 coor.test <- function(X = NULL, iterations = 1000, less.than.chance = TRUE, parallel = 1, pb = TRUE, 
                       rm.imcomp = FALSE, cutoff = 2, rm.solo = FALSE)
@@ -167,7 +183,7 @@ if(any(!cse)) warning("Some events didn't have 2 individuals and were excluded")
   obs.overlaps <- length(ovlp[ovlp=="ovlp"])
   mean.random.ovlps <- mean(rov)
   if(less.than.chance) p <- length(rov[rov <= obs.overlaps])/iterations else p <- length(rov[rov >= obs.overlaps])/iterations
-  l <- data.frame(h, obs.overlaps, mean.random.ovlps, p)
+  l <- data.frame(sing.event = h, obs.ovlps = obs.overlaps, mean.random.ovlps, p.value = p, coor.score = round((obs.overlaps - mean.random.ovlps)/mean.random.ovlps, 3))
   
   return(l)}
       # )
@@ -187,7 +203,5 @@ if(any(!cse)) warning("Some events didn't have 2 individuals and were excluded")
     
   df <- do.call(rbind, cote)
     
-colnames(df) <- c("sing.event", "obs.ovlps", "mean.random.ovlps", "p.value")
-
 return(df)
 }