+++ R/read_wave.R
@@ -1,10 +1,44 @@
-#' A wrapper for tuner readWave function internal warbleR function not to be called by users
-# \code{read_wave} A wrapper for tuneR's \code{\link[tuneR]{readWave}} function that deals with extension case mismatches \code{\link{manualoc}}) in 
-#' which the start/end times and frequency range of acoustic signals listed in a data frame can be adjusted.
-# @author Marcelo Araya-Salas (\email{araya-salas@@cornell.edu})
-#last modification on jul-5-2016 (MAS)
+#' A wrapper for tuneR's readWave 
+#' 
+#' \code{read_wave} A wrapper for tuneR's \code{\link[tuneR]{readWave}} function that deals with extension case mismatches.
+#' @usage read_wave(filename, from = 0, to = Inf, units = "seconds", header = FALSE,
+#' toWaveMC = NULL, ignore.ext.case = TRUE) 
+#' @param filename Filename of the file to be read.
+#' @param from Where to start reading (in order to save memory by reading wave file piecewise), in units. 
+#' @param to Where to stop reading (in order to save memory by reading wave file piecewise), in units. Default is \code{Inf}.
+#' @param units	Units in which from and to is given, the default is "seconds".
+#' @param header If \code{TRUE}, just header information of the Wave file are returned, otherwise (the default) the whole Wave object.
+#' @param toWaveMC If \code{TRUE}, a WaveMC-class object is returned. If \code{NULL}
+#' (default) or \code{FALSE} and a non-extensible Wave file or an extensible Wave
+#'  file with no other than the 'FL' and 'FR' channels is found, a Wave-class object
+#'  is returned, otherwise a WaveMC-class object.
+#' @param ignore.ext.case Logical to control if the file extension case is ignored. Default is \code{TRUE}.
+#' @return An object of class "Wave".
+#' @export
+#' @name read_wave
+#' @details The function is a wrapper for \code{\link[tuneR]{readWave}} that ignores file extension mismatches, a common mistake when reading wave files.
+#' @examples
+#' {
+#' # First set temporary folder
+#' # setwd(tempdir())
+#' 
+#' # write wave files with lower case file extension
+#' data(list = c("Phae.long1"))
+#' writeWave(Phae.long1,"Phae.long1.wav")
+#' 
+#' # provide right extension case
+#' read_wave("Phae.long1.wav")
+#' 
+#' # provide wrong extension case
+#' read_wave("Phae.long1.WAV")
+#' 
+#' #' # provide no extension
+#' read_wave("Phae.long1")
+#' }
+#' @author Marcelo Araya-Salas (\email{araya-salas@@cornell.edu})
+#last modification on apr-18-2018 (MAS)
 
-read_wave <- function (filename, from = 1, to = Inf, units = "seconds", header = FALSE, toWaveMC = NULL, ignore.ext.case = TRUE) 
+read_wave <- function (filename, from = 0, to = Inf, units = "seconds", header = FALSE, toWaveMC = NULL, ignore.ext.case = TRUE) 
 {
   if (length(filename) != 1) 
     stop("Please specify exactly one 'filename'.")
@@ -17,15 +51,15 @@ read_wave <- function (filename, from = 1, to = Inf, units = "seconds", header =
   if(ignore.ext.case)
   {
     # check if is actually a file extension
-  if (extnt %in% c(".wav", ".WAV"))  
-    no.ext.name <- substr(x = filename, start = 0, stop = nchar(filename) - 4) else
-      no.ext.name <- filename
-    
-  # paste file name and file extension
-    if (file.exists(paste0(no.ext.name, ".wav"))) filename <- paste0(no.ext.name, ".wav")  else
-      if (file.exists(paste0(no.ext.name, ".WAV"))) filename <- paste0(no.ext.name, ".wav")  
-}
-   if (!file.exists(filename)) 
+    if (extnt %in% c(".wav", ".WAV"))  
+      no.ext.name <- substr(x = filename, start = 0, stop = nchar(filename) - 4) else
+        no.ext.name <- filename
+      
+      # paste file name and file extension
+      if (file.exists(paste0(no.ext.name, ".wav"))) filename <- paste0(no.ext.name, ".wav")  else
+        if (file.exists(paste0(no.ext.name, ".WAV"))) filename <- paste0(no.ext.name, ".wav")  
+  }
+  if (!file.exists(filename)) 
     stop("File '", filename, "' does not exist.")
   
   if (file.access(filename, 4)) 
@@ -152,7 +186,7 @@ read_wave <- function (filename, from = 1, to = Inf, units = "seconds", header =
     isTRUE(toWaveMC)
   else TRUE
   if (toWaveMC) {
-    object <- new("WaveMC", samp.rate = sample.rate, bit = bits, 
+    object <- methods::new("WaveMC", samp.rate = sample.rate, bit = bits, 
                   pcm = !(pcm == 3 || (exists("SubFormat") && SubFormat == 
                                          3)))
     object@.Data <- matrix(sample.data, ncol = channels,