+++ R/warbleR_options.R
@@ -1,0 +1,93 @@
+#' Setting warbleR options 
+#' 
+#' \code{warbleR_options} sets parameters for warbleR functions
+#' @usage warbleR_options(...)
+#' @param ... Arguments in `parameter = value`` form, or a list of tagged values. The tags (i.e. parameters)
+#'  must come from the list of parameters described below.
+#' @return When parameters are set by warbleR_options, their former values are 
+#' returned in an invisible named list. Such a list can be passed as an argument to 
+#' pboptions to restore the parameter values. IF the function is called with no arguments the current option values are printed.
+#' @export
+#' @name warbleR_options
+#' @details The function aims to simplify the use of global of parameters that apply to many warbleR functions 
+#' by setting a default value that will be used to any function in downstream 
+#' analyses. All values are set to \code{NULL} by 
+#' default. Tags that are set with warbleR_options will be used iby the functions 
+#' that share those arguments. However, if an argument is set within a function call
+#' it will overwrite the values set by warbleR_options. The following tags are available: 
+#' \itemize{
+#'    \item \code{bp}: Numeric vector of length 2 giving the lower and upper limits of a
+#'   frequency bandpass filter (in kHz).
+#'    \item \code{collevels}:  A numeric vector of length 3. Specifies levels to partition the 
+#'   amplitude range of the spectrogram (in dB). The more levels the higher the
+#'   resolution of the spectrogram. THe lower the first value the darker the spectrograms.
+#'    \item \code{dest.path}: Character string containing the directory path where 
+#'    the cut sound files will be saved. Used in \code{\link{consolidate}} and \code{\link{cut_sels}}.
+#'    \item \code{flim}: A numeric vector of length 2 for the frequency limit in kHz of 
+#'   the spectrogram, as in \code{\link[seewave]{spectro}}.
+#'    \item \code{img.path}: Character string containing the directory path where the images should be saved. Used as 'path' in all image creating functions.
+#'    \item \code{it}: A character vector of length 1 giving the image type to be used. Currently only
+#' "tiff" and "jpeg" are admitted.
+#'    \item \code{osci}: Logical argument to add an oscillogram underneath spectrogram, as
+#'   in \code{\link[seewave]{spectro}}.
+#'    \item \code{pal}: A color palette function to be used to assign colors in the 
+#'   plot, as in \code{\link[seewave]{spectro}}.
+#'    \item \code{parallel}: Numeric. Controls whether parallel computing is applied.
+#' It specifies the number of cores to be used in iterative functions. 
+#'    \item \code{pb}: Logical argument to control whether progress bar is used. 
+#'    \item \code{res}: Numeric argument of length 1. Controls image resolution in all image creating functions.
+#'    \item \code{wav.path}: Character string containing the directory path where the
+#'    sound files are located. Used as 'path' in all functions in which sound files are read.
+#'    \item \code{wl}: A numeric vector of length 1 specifying the window length for creating spectrogram (either for plotting or for measuring spectrogram parameters).
+#'    \item \code{wn}: Character vector of length 1 specifying the window name for creating spectrogram (either for plotting or for measuring spectrogram parameters). See function \code{\link[seewave]{ftwindow}} for options.
+#'    }
+#' @examples
+#' {
+#' # First set temporary folder
+#' # setwd(tempdir())
+#' 
+#' data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))
+#' writeWave(Phae.long1,"Phae.long1.wav")
+#' writeWave(Phae.long2,"Phae.long2.wav")
+#' writeWave(Phae.long3,"Phae.long3.wav")
+#' writeWave(Phae.long4,"Phae.long4.wav")
+#' 
+#' # sig2noise with progress bar (by default is TRUE)
+#' a <- sig2noise(X = selec.table, mar = 0.1)
+#' 
+#' # set progress bar to FALSE with warbleR_options
+#' warbleR_options(pb = FALSE)
+#' 
+#' # sig2noise without progress bar
+#' a <- sig2noise(X = selec.table, mar = 0.1)
+#' 
+#' # sig2noise with progress bar by setting in the function call (overwritting options)
+#' a <- sig2noise(X = selec.table, pb = TRUE, mar = 0.1)
+#' }
+#' @author Marcelo Araya-Salas (\email{araya-salas@@cornell.edu})
+# last modification on apr-18-2018 (MAS)
+
+warbleR_options <- function(...){
+  opar <- getOption("warbleR")
+  argms <- list(...)
+  
+  if (length(argms) > 0) {
+    if (!is.null(argms$wav.path)) if (!dir.exists(argms$wav.path)) stop("provided 'wav.path' doesn't exist")
+    
+    if (!is.null(argms$img.path)) if (!dir.exists(argms$img.path)) stop("provided 'dest.path' doesn't exist")
+    
+    if (!is.null(argms$dest.path)) if (!dir.exists(argms$dest.path)) stop("provided 'dest.path' doesn't exist")
+    
+    if (length(argms)) {
+      if (length(argms) == 1 && is.list(argms[[1]])) {
+        npar <- argms[[1]]
+      } else {
+        npar <- opar
+        npar[match(names(argms), names(npar))] <- argms
+      }
+      options("warbleR" = npar)
+    }
+    invisible(opar)
+  } else print(opar)
+  
+}